% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_summaries.R
\name{all_summaries_by_year}
\alias{all_summaries_by_year}
\title{Scrape All School Statistics and Summaries for Given Year}
\usage{
all_summaries_by_year(year)
}
\arguments{
\item{year}{Year for which to scrape all school statistics and summaries for
schools in Division I}
}
\value{
Data Frame with all school statistics and summaries for given year
}
\description{
Scrape All School Statistics and Summaries for Given Year
}
