% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{AST}
\alias{AST}
\title{Calculate Assist Percentage}
\usage{
AST(TM_AST, TM_FG)
}
\arguments{
\item{TM_AST}{One team's assists}

\item{TM_FG}{One team's field goals made}
}
\value{

}
\description{
Calculate Assist Percentage
}
