% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banks}
\alias{banks}
\title{Data for DD analysis of the effect of monetary policy on bank failures in Mississippi in 1930}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1878 rows and 9 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
banks
}
\description{
Data for DD analysis of the effect of monetary policy on bank failures in Mississippi in 1930
}
\keyword{datasets}

