% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{BLK}
\alias{BLK}
\title{Calculate Block Percentage}
\usage{
BLK(TM_BLK, poss)
}
\arguments{
\item{TM_BLK}{One team's blocks}

\item{poss}{Number of possessions in game}
}
\value{

}
\description{
Calculate Block Percentage
}
