% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{child_labor_law}
\alias{child_labor_law}
\title{Data for analysis of the returns to schooling using child labor laws as instruments}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 722343 rows and 25 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
child_labor_law
}
\description{
Data for analysis of the returns to schooling using child labor laws as instruments
}
\keyword{datasets}

