% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd_mlda}
\alias{dd_mlda}
\title{Data for DD analysis of the minimum legal drinking age}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24786 rows and 15 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
dd_mlda
}
\description{
Data for DD analysis of the minimum legal drinking age
}
\keyword{datasets}

