% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{DRB}
\alias{DRB}
\title{Calculate Defensive Rebounding Percentage}
\usage{
DRB(TM_DRB, O_ORB)
}
\arguments{
\item{TM_DRB}{One team's defensive rebounds}

\item{O_ORB}{Opponent's offensive rebounds}
}
\value{

}
\description{
Calculate Defensive Rebounding Percentage
}
