% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{eFG}
\alias{eFG}
\title{Calculate Effective Field Goal Percentage}
\usage{
eFG(TM_FG, TM_3PM, TM_FGA)
}
\arguments{
\item{TM_FG}{One team's field goals made}

\item{TM_3PM}{One team's made 3-pointers}

\item{TM_FGA}{One team's field goals attempted}
}
\value{

}
\description{
Calculate Effective Field Goal Percentage
}
