% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{FTr}
\alias{FTr}
\title{Calculate Free Throw Rate}
\usage{
FTr(TM_FTA, TM_FGA)
}
\arguments{
\item{TM_FTA}{One team's free throws attempted}

\item{TM_FGA}{One team's field goals attempted}
}
\value{

}
\description{
Calculate Free Throw Rate
}
