% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_diploma}
\alias{hs_diploma}
\title{Data for analysis of the “sheepskin effects” of a high school diploma}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 5 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
hs_diploma
}
\description{
Data for analysis of the “sheepskin effects” of a high school diploma
}
\keyword{datasets}

