% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdve}
\alias{mdve}
\title{The Minneapolis Domestic Violence Experiment (MDVE)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 330 rows and 27 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
mdve
}
\description{
The Minneapolis Domestic Violence Experiment (MDVE)
}
\keyword{datasets}

