% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhis2009}
\alias{nhis2009}
\title{2009 National Health Interview Survey (NHIS)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80634 rows and 40 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
nhis2009
}
\description{
2009 National Health Interview Survey (NHIS)
}
\keyword{datasets}

