% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{ORB}
\alias{ORB}
\title{Calculate Offensive Rebounding Percentage}
\usage{
ORB(TM_ORB, O_DRB)
}
\arguments{
\item{TM_ORB}{One team's offensive rebounds}

\item{O_DRB}{Opponent's defensive rebounds}
}
\value{

}
\description{
Calculate Offensive Rebounding Percentage
}
