% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{poss}
\alias{poss}
\title{Estimate possessions in a game}
\usage{
poss(
  TM_FGA,
  TM_FTA,
  TM_ORB,
  O_TRB,
  O_ORB,
  TM_FG,
  TM_TOV,
  O_FGA,
  O_FTA,
  TM_TRB,
  O_FG,
  O_TOV
)
}
\arguments{
\item{TM_FGA}{One team's field goals attempted}

\item{TM_FTA}{One team's free throws attempted}

\item{TM_ORB}{One team's offensive rebounds}

\item{O_TRB}{Opponent's total rebounds}

\item{O_ORB}{Opponent's offensive rebounds}

\item{TM_FG}{One team's field goals made}

\item{TM_TOV}{One team's turnovers}

\item{O_FGA}{Opponent's field goals attempted}

\item{O_FTA}{Opponent's free throws attempted}

\item{TM_TRB}{One team's total rebounds}

\item{O_FG}{Opponent's field goals made}

\item{O_TOV}{Opponent's turnovers}
}
\value{

}
\description{
Estimate possessions in a game
}
