% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qob}
\alias{qob}
\title{Data for analysis of the returns to schooling using quarters of birth (QOB) as instruments}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 329509 rows and 5 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
qob
}
\description{
Data for analysis of the returns to schooling using quarters of birth (QOB) as instruments
}
\keyword{datasets}

