% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rand_hie_annual_spend}
\alias{rand_hie_annual_spend}
\title{The RAND Health Insurance Experiment (HIE)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25740 rows and 8 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
rand_hie_annual_spend
}
\description{
The RAND Health Insurance Experiment (HIE): rand_hie_annual_spend
}
\keyword{datasets}

