% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rd_mlda}
\alias{rd_mlda}
\title{Data for RD analysis of the minimum legal drinking age}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 19 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
rd_mlda
}
\description{
Data for RD analysis of the minimum legal drinking age
}
\keyword{datasets}

