% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_summaries.R
\name{school_summary_by_year}
\alias{school_summary_by_year}
\title{Scrape Various Statistics and Summaries by Team and Year}
\usage{
school_summary_by_year(team_id, year)
}
\arguments{
\item{team_id}{Team ID for team of interest from teams table}

\item{year}{Year to pull data from}
}
\value{
Data Frame with a variety of statistics and summaries by team and year
}
\description{
Scrape Various Statistics and Summaries by Team and Year
}
