% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preseason_rankings.R
\name{scrape_preseason_odds}
\alias{scrape_preseason_odds}
\title{Scrape preseason odds}
\usage{
scrape_preseason_odds(year)
}
\arguments{
\item{year}{Year for which to scrape preseason odds}
}
\value{
Data frame with all team's preseason odds for given year
}
\description{
Scrape preseason odds
}
