% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preseason_rankings.R
\name{scrape_preseason_ranks}
\alias{scrape_preseason_ranks}
\title{Scraping preseason top 25 rankings}
\usage{
scrape_preseason_ranks(year, table = 1)
}
\arguments{
\item{year}{Year for which to scrape preseason top 25 rankings}

\item{table}{Table number to scrape. Should be 1 for all but the current season}
}
\value{
Data frame with top 25 teams
}
\description{
Scraping preseason top 25 rankings
}
