% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_summaries.R
\name{scrape_teams_table}
\alias{scrape_teams_table}
\title{Scrape Teams Table from basketball-reference}
\usage{
scrape_teams_table()
}
\value{
Data frame with all Division I NCAA men's basketball teams, their
years in Division I, and the team IDs of each team
}
\description{
Scrape Teams Table from basketball-reference
}
