% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{STL}
\alias{STL}
\title{Calculate Steal Percentage}
\usage{
STL(TM_STL, poss)
}
\arguments{
\item{TM_STL}{One team's steals}

\item{poss}{Number of possessions in game}
}
\value{

}
\description{
Calculate Steal Percentage
}
