% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{TO}
\alias{TO}
\title{Calculate Turnover Percentage}
\usage{
TO(TM_TOV, poss)
}
\arguments{
\item{TM_TOV}{One team's turnovers}

\item{poss}{Number of possessions in game}
}
\value{

}
\description{
Calculate Turnover Percentage
}
