% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{TPr}
\alias{TPr}
\title{Calculate Three Point Rate}
\usage{
TPr(TM_3P, TM_FGA)
}
\arguments{
\item{TM_3P}{One team's 3-pointers made}

\item{TM_FGA}{One team's field goals attempted}
}
\value{

}
\description{
Calculate Three Point Rate
}
