% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{TRB}
\alias{TRB}
\title{Calculate Total Rebounding Percentage}
\usage{
TRB(TM_TRB, O_TRB)
}
\arguments{
\item{TM_TRB}{One team's total rebounds}

\item{O_TRB}{Opponent's total rebounds}
}
\value{

}
\description{
Calculate Total Rebounding Percentage
}
