% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_stats.R
\name{TS}
\alias{TS}
\title{Calculate True Shooting Percentage}
\usage{
TS(TM_PTS, TM_FGA, TM_FTA)
}
\arguments{
\item{TM_PTS}{One team's total points scored}

\item{TM_FGA}{One team's field goals attempted}

\item{TM_FTA}{One team's free throws attempted}
}
\value{

}
\description{
Calculate True Shooting Percentage
}
