% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twins}
\alias{twins}
\title{Data for analysis of the returns to schooling for Twinsburg twins}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 680 rows and 39 columns.}
\source{
\url{http://masteringmetrics.com/resources/}
}
\usage{
twins
}
\description{
Data for analysis of the returns to schooling for Twinsburg twins
}
\keyword{datasets}

